Add `IUserAPIClient` to make swapping `UserManager` easier.
This harmonizes implementation of `UserManager` with other
managers both in terms of API and code organisation.

Copy-paste errors in docstring of `SERVICE_USER_URL` and of polls
(incorrectly referencing "workspaces" and "specs" respectively)
were corrected.
